extract file uploads into separate hook
